Add ios/configure script
The script performs a ./configure for each iOS architecture ("slice")
and stores the result in the ios/build/ARCH directories.  It then
creates a Makefile that invokes all the configured Makefiles to build a
separate libservaldaemon.a for each architecture, then invokes the
lipo(1) utility to merge them into a single, multi-slice library.

This enables normal development on iOS using the standard edit-make-test
cycle, instead of having to re-configure and re-compile for each
architecture after every edit.

Uses a feature of the autoconf ./configure script: if invoked from a
current working directory that is not the Serval DNA repo root, then it
will write its output files (Makefile, config.h, config.status etc.)
into the current working directory, not into the repo root.

Updates README.md and adds a technical doc for iOS (empty for now).