Add Swift 3 Keyring API
Adds a swift3-api subdirectory containing the Swift 3 source code and a
Makefile.in that compiles it into the "ServalClient" Swift package using
the Swift 3 package manager.

ServalKeyring supports the operations: add, remove, set, list.  The
Swift code for AbstractId, SubscriberId is already in near-final form.
ServalRestfulClient and ServalKeyring are still work-in-progress.

Adds autoconf macros for the Swift compiler and package manager.